+++ src/data_inclusion/schema/modes_orientation.py
@@ -5,27 +5,27 @@ class ModeOrientationAccompagnateur(EnhancedEnum):
     COMPLETER_LE_FORMULAIRE_DADHESION = (
         "completer-le-formulaire-dadhesion",
         "Compléter le formulaire d’adhésion",
-        None,
+        "Compléter le formulaire d’adhésion associé.",
     )
     ENVOYER_UN_MAIL = (
         "envoyer-un-mail",
-        "Envoyer un mail",
-        None,
+        "Envoyer un courriel",
+        "Envoyer un courriel à l’adresse indiquée.",
     )
     ENVOYER_UN_MAIL_AVEC_UNE_FICHE_DE_PRESCRIPTION = (
         "envoyer-un-mail-avec-une-fiche-de-prescription",
-        "Envoyer un mail avec une fiche de prescription",
-        None,
+        "Envoyer un courriel avec une fiche de prescription",
+        "Envoyer un courriel à l’adresse indiquée avec une fiche de prescription.",
     )
     ENVOYER_UN_MAIL_AVEC_DES_DOCUMENTS_A_COMPLETER = (
         "envoyer-un-mail-avec-des-documents-a-completer",
-        "Envoyer un mail avec des documents à compléter",
-        None,
+        "Envoyer un courriel avec des documents à compléter",
+        "Envoyer un courriel à l’adresse indiquée avec des documents à compléter.",
     )
     TELEPHONER = (
         "telephoner",
         "Téléphoner",
-        None,
+        "Téléphoner au numéro indiqué.",
     )
     AUTRE = (
         "autre",
@@ -38,22 +38,22 @@ class ModeOrientationBeneficiaire(EnhancedEnum):
     COMPLETER_LE_FORMULAIRE_DADHESION = (
         "completer-le-formulaire-dadhesion",
         "Compléter le formulaire d’adhésion",
-        None,
+        "Compléter le formulaire d’adhésion associé.",
     )
     ENVOYER_UN_MAIL = (
         "envoyer-un-mail",
-        "Envoyer un mail",
-        None,
+        "Envoyer un courriel",
+        "Envoyer un courriel à l’adresse indiquée.",
     )
     SE_PRESENTER = (
         "se-presenter",
         "Se présenter",
-        None,
+        "Se présenter à l’adresse indiquée.",
     )
     TELEPHONER = (
         "telephoner",
         "Téléphoner",
-        None,
+        "Téléphoner au numéro indiqué.",
     )
     AUTRE = (
         "autre",